+++ src/sieve_kernel.cuh
@@ -13,7 +13,7 @@ __global__ void segmented_sieve_kernel(
     uint32_t tiny_prime_count,
     uint32_t small_prime_count,
     const uint64_t* __restrict__ bucket_hits,
-    const uint32_t* __restrict__ bucket_offsets,
+    const uint64_t* __restrict__ bucket_offsets,
     uint64_t* __restrict__ output_primes,
     uint32_t max_output,
     uint32_t* output_count
